requirements.txt
flask
requests
beautifulsoup4
transformers
torch
gtts
streamlit
plotly
vaderSentiment
sentencepiece
lxml==4.9.3
newspaper3k==0.2.8